# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Backlogg API",
    "description": "Backlogg is a unified catalog API for **movies, series, books and games**,\nseeded from public external APIs (TMDB, Open Library, IGDB) via a nightly sync,\nplus a social layer on top of that catalog.\n\n**Catalog & discovery:** list and detail endpoints per content type, cross-type\n`/search` (with on-demand external fallback), `/genres`, `/trending`, `/people`\nand similar-item recommendations.\n\n**Social layer:** account management with refresh tokens and account recovery\n(`/auth`, `/users`), ratings & reviews, per-user library/backlog, curated\n`/lists`, `/follows`, an activity `/feed`, `/notifications` and personalized\n`/recommendations`.\n\n**Platform:** rate limiting, structured observability, Prometheus `/metrics`\nand response caching.\n\n## Authentication\n\n- **User endpoints** use `Authorization: Bearer <access_token>`. Obtain the\n  token pair from `POST /auth/login`; renew it with `POST /auth/refresh`.\n- **Admin endpoints** (`/admin/*`) require the `X-API-Key` header.\n\nEverything else is public. Direct messaging between users is out of scope.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/notifications": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "List notifications",
        "description": "The caller's notifications, paginated, newest first. Requires auth.",
        "operationId": "list_notifications_v1_notifications_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Limit"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/notifications/unread_count": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "Unread notification count",
        "description": "Number of unread notifications for the caller. Requires auth.",
        "operationId": "unread_count_v1_notifications_unread_count_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnreadCountOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/notifications/read": {
      "post": {
        "tags": [
          "notifications"
        ],
        "summary": "Mark notifications read",
        "description": "Mark notifications read. Empty body marks all, else only given ids. Idempotent.",
        "operationId": "mark_read_v1_notifications_read_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/MarkReadIn"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/notifications/{notification_id}": {
      "delete": {
        "tags": [
          "notifications"
        ],
        "summary": "Delete a notification",
        "description": "Delete one of the caller's own notifications. Requires auth. 404 if it doesn't exist or doesn't belong to the caller.",
        "operationId": "delete_notification_v1_notifications__notification_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Notification Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MarkReadIn": {
        "properties": {
          "ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ids"
          }
        },
        "type": "object",
        "title": "MarkReadIn",
        "description": "Body for POST /notifications/read.\n\n``ids`` omitted / null \u2192 mark all unread as read. ``ids`` provided \u2192 mark\nonly those (idempotent)."
      },
...
      "NotificationActorOut": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          }
        },
        "type": "object",
        "required": [
          "username",
          "display_name",
          "avatar_url"
        ],
        "title": "NotificationActorOut",
        "description": "The user who triggered the notification."
      },
      "NotificationListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NotificationOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "limit"
        ],
        "title": "NotificationListOut",
        "example": {
          "items": [
            {
              "actor": {
                "avatar_url": "https://cdn.example.com/a/bob.png",
                "display_name": "Bob",
                "username": "bob"
              },
              "created_at": "2026-05-25T18:04:11Z",
              "id": 91,
              "is_read": false,
              "target": {
                "item_type": "MOVIE",
                "slug": "the-matrix-1999",
                "target_id": 512,
                "target_type": "review"
              },
              "type": "review_like"
            }
          ],
          "limit": 20,
          "page": 1,
          "total": 5
        }
      },
      "NotificationOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "actor": {
            "$ref": "#/components/schemas/NotificationActorOut"
          },
          "target": {
            "$ref": "#/components/schemas/NotificationTargetOut"
          },
          "is_read": {
            "type": "boolean",
            "title": "Is Read"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "actor",
          "target",
          "is_read",
          "created_at"
        ],
        "title": "NotificationOut"
      },
      "NotificationTargetOut": {
        "properties": {
          "target_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Type"
          },
          "target_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Id"
          },
          "item_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Type"
          },
          "slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          }
        },
        "type": "object",
        "required": [
          "target_type",
          "target_id",
          "item_type",
          "slug"
        ],
        "title": "NotificationTargetOut",
        "description": "Polymorphic target of the notification (null for new_follower).\n\n``item_type``/``slug`` resolve the review's target item (movie/series/\nbook/game) so a client can link directly to it without a separate\n``GET /v1/ratings/{id}`` lookup \u2014 there is none. Populated only when\n``target_type == \"review\"``; ``null`` otherwise (including\n``new_follower``). ``item_type`` is uppercase (``\"MOVIE\"``, ``\"SERIES\"``,\n``\"BOOK\"``, ``\"GAME\"``) \u2014 same convention as the feed endpoints."
      },
...
      "UnreadCountOut": {
        "properties": {
          "unread_count": {
            "type": "integer",
            "title": "Unread Count"
          }
        },
        "type": "object",
        "required": [
          "unread_count"
        ],
        "title": "UnreadCountOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}